Even more optimizations: Added it so you can load in the existing file and edit the stage from there, also made checks so you had to at least attempt to load in a file before being able to edit it (to prevent writing over a level that someone made).
Also with the changes the stage will now be able to easily be built through a text file horizontaly and will allow us to make a stage as tall or as short as we want. We also do not need format characters as well.

Problems: With this change it means that every time you load in the existing file the title of the file and the level number loads in the text box at the top (which means whoever is making the level will have to manually delete this before compiling changes). Also when reloading the text file into the textbox and extra line is created at the bottom of the text box, I have yet to figure out how to get rid of these problems.