Added a character to advance y-axis on the Stage Builder and Stage Reader (Stage should now load correctly if using the new Test file). Rewrote some of the StageBuilder with xpos and ypos variables (instead of just x) also got multiple lines working. Got the player to spawn correctly in the built stage. Changed it so the right player texture spawns, not the green cheese looking thing...
What still needs to still be done:
The Stage Reader's reading path needs to be changed to the content bin for "blank-canvas" - this is where the levels will be when game is finished. Also it needs to be changed to read a specific text file in the bin for when we load specific levels later on.
Physics seem weird when you get the character on the stage. The player continuously gyrates, may want to fix it.